02-application/requirements.txt
kafka-python==2.0.2
Flask==3.0.3
